+++ tests/client.rs
@@ -642,3 +642,24 @@ fn can_get_build_with_alias() {
     assert_that!(jenkins.get_build("normal job", "lastCompletedBuild")).is_ok();
     assert_that!(jenkins.get_build("normal job", "zut")).is_err();
 }
+
+#[test]
+fn can_get_nodes() {
+    setup();
+    let jenkins = JenkinsBuilder::new(JENKINS_URL)
+        .with_user("user", Some("password"))
+        .build()
+        .unwrap();
+    assert_that!(jenkins.get_nodes()).is_ok();
+}
+
+#[test]
+fn can_get_master() {
+    setup();
+    let jenkins = JenkinsBuilder::new(JENKINS_URL)
+        .with_user("user", Some("password"))
+        .build()
+        .unwrap();
+    assert_that!(jenkins.get_node("(master)")).is_ok();
+    assert_that!(jenkins.get_master_node()).is_ok();
+}